third_party/WebKit/LayoutTests/http/tests/loading/module-script-wrong-mime.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script type="module">
import "./resources/module-script-wrong-mime.txt";
</script>
<script type="module">
test(
  () => assert_equals(undefined, window.ran_invalid_mime_script),
  "module script with invalid mime type shouldn't have run.");
</script>
